refactor(in-app): extract common dialog logic to BaseDialog
