+++ REES46/Classes/InAppNotification/view/component/dialog/types/BottomDialog.swift
@@ -1,0 +1,32 @@
+import UIKit
+
+class BottomDialog: BaseDialog {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = UIColor.black.withAlphaComponent(0.5)
+    }
+    
+    override func setupContentView() {
+        super.setupContentView()
+        contentView.layer.cornerRadius = AppDimensions.Padding.medium
+        contentView.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
+    }
+    
+    override func setContentContainerConstraints() {
+        NSLayoutConstraint.activate([
+            contentContainer.topAnchor.constraint(equalTo: imageContainer.isHidden ? contentView.topAnchor : imageContainer.bottomAnchor),
+            contentContainer.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            contentContainer.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            contentContainer.bottomAnchor.constraint(equalTo: contentView.safeAreaLayoutGuide.bottomAnchor) 
+        ])
+    }
+    
+    override func setContentViewConstraints() {
+        NSLayoutConstraint.activate([
+            contentView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            contentView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            contentView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+        ])
+    }
+}
