+++ REES46/Classes/InAppNotification/view/component/dialog/types/FullScreenDialog.swift
