+++ REES46/Classes/InAppNotification/view/component/dialog/utils/Dialog.constraints.swift
@@ -1,0 +1,119 @@
+import UIKit
+
+extension BaseDialog {
+    func applyConstraints() {
+        setContentViewConstraints()
+        setImageContainerConstraints()
+        setBackgroundImageViewConstraints()
+        setContentContainerConstraints()
+        setCloseButtonConstraints()
+        setTitleLabelConstraints()
+        setMessageLabelConstraints()
+        setButtonConstraints(buttonState: determineButtonState())
+    }
+    
+    private func setImageContainerConstraints() {
+        if imageUrl.isEmpty {
+            NSLayoutConstraint.activate([
+                imageContainer.topAnchor.constraint(equalTo: contentView.topAnchor),
+                imageContainer.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+                imageContainer.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+                imageContainer.heightAnchor.constraint(equalToConstant: 0)
+            ])
+        } else {
+            NSLayoutConstraint.activate([
+                imageContainer.topAnchor.constraint(equalTo: contentView.topAnchor),
+                imageContainer.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+                imageContainer.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+                imageContainer.heightAnchor.constraint(equalToConstant: AppDimensions.Size.bottomSheetHeight)
+            ])
+        }
+    }
+    
+    private func setBackgroundImageViewConstraints() {
+        NSLayoutConstraint.activate([
+            backgroundImageView.topAnchor.constraint(equalTo: imageContainer.topAnchor),
+            backgroundImageView.leadingAnchor.constraint(equalTo: imageContainer.leadingAnchor),
+            backgroundImageView.trailingAnchor.constraint(equalTo: imageContainer.trailingAnchor),
+            backgroundImageView.bottomAnchor.constraint(equalTo: imageContainer.bottomAnchor)
+        ])
+    }
+    
+    private func setCloseButtonConstraints() {
+        NSLayoutConstraint.activate([
+            closeButton.topAnchor.constraint(equalTo: imageContainer.topAnchor, constant: AppDimensions.Padding.small),
+            closeButton.trailingAnchor.constraint(equalTo: imageContainer.trailingAnchor, constant: -AppDimensions.Padding.small),
+            closeButton.widthAnchor.constraint(equalToConstant: AppDimensions.Size.closeButtonSize),
+            closeButton.heightAnchor.constraint(equalToConstant: AppDimensions.Size.closeButtonSize)
+        ])
+    }
+    
+    private func setTitleLabelConstraints() {
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: contentContainer.topAnchor, constant: AppDimensions.Padding.medium),
+            titleLabel.leadingAnchor.constraint(equalTo: contentContainer.leadingAnchor, constant: AppDimensions.Padding.medium),
+            titleLabel.trailingAnchor.constraint(equalTo: contentContainer.trailingAnchor, constant: -AppDimensions.Padding.medium)
+        ])
+    }
+    
+    private func setMessageLabelConstraints() {
+        NSLayoutConstraint.activate([
+            messageLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: AppDimensions.Padding.small),
+            messageLabel.leadingAnchor.constraint(equalTo: contentContainer.leadingAnchor, constant: AppDimensions.Padding.medium),
+            messageLabel.trailingAnchor.constraint(equalTo: contentContainer.trailingAnchor, constant: -AppDimensions.Padding.medium)
+        ])
+    }
+    
+    private func setButtonConstraints(buttonState: ButtonState) {
+        switch buttonState {
+            case .noButtons:
+                NSLayoutConstraint.activate([
+                    messageLabel.bottomAnchor.constraint(equalTo: contentContainer.bottomAnchor, constant: -AppDimensions.Padding.medium)
+                ])
+                
+            case .onlyAccept:
+                NSLayoutConstraint.activate([
+                    confirmButton.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: AppDimensions.Padding.medium),
+                    confirmButton.leadingAnchor.constraint(equalTo: contentContainer.leadingAnchor, constant: AppDimensions.Padding.medium),
+                    confirmButton.trailingAnchor.constraint(equalTo: contentContainer.trailingAnchor, constant: -AppDimensions.Padding.medium),
+                    confirmButton.heightAnchor.constraint(equalToConstant: AppDimensions.Height.popUpButton),
+                    confirmButton.bottomAnchor.constraint(equalTo: contentContainer.bottomAnchor, constant: -AppDimensions.Padding.medium)
+                ])
+                
+            case .onlyDecline:
+                NSLayoutConstraint.activate([
+                    dismissButton.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: AppDimensions.Padding.medium),
+                    dismissButton.leadingAnchor.constraint(equalTo: contentContainer.leadingAnchor, constant: AppDimensions.Padding.medium),
+                    dismissButton.trailingAnchor.constraint(equalTo: contentContainer.trailingAnchor, constant: -AppDimensions.Padding.medium),
+                    dismissButton.heightAnchor.constraint(equalToConstant: AppDimensions.Height.popUpButton),
+                    dismissButton.bottomAnchor.constraint(equalTo: contentContainer.bottomAnchor, constant: -AppDimensions.Padding.medium)
+                ])
+                
+            case .bothButtons:
+                NSLayoutConstraint.activate([
+                    dismissButton.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: AppDimensions.Padding.medium),
+                    dismissButton.leadingAnchor.constraint(equalTo: contentContainer.leadingAnchor, constant: AppDimensions.Padding.medium),
+                    dismissButton.trailingAnchor.constraint(equalTo: contentContainer.centerXAnchor, constant: -AppDimensions.Padding.small),
+                    dismissButton.heightAnchor.constraint(equalToConstant: AppDimensions.Height.popUpButton),
+                    
+                    confirmButton.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: AppDimensions.Padding.medium),
+                    confirmButton.leadingAnchor.constraint(equalTo: contentContainer.centerXAnchor, constant: AppDimensions.Padding.small),
+                    confirmButton.trailingAnchor.constraint(equalTo: contentContainer.trailingAnchor, constant: -AppDimensions.Padding.medium),
+                    confirmButton.heightAnchor.constraint(equalToConstant: AppDimensions.Height.popUpButton),
+                    confirmButton.bottomAnchor.constraint(equalTo: contentContainer.bottomAnchor, constant: -AppDimensions.Padding.medium)
+                ])
+        }
+    }
+    
+    private func determineButtonState() -> ButtonState {
+        if confirmButton.isHidden && dismissButton.isHidden {
+            return .noButtons
+        } else if confirmButton.isHidden {
+            return .onlyDecline
+        } else if dismissButton.isHidden {
+            return .onlyAccept
+        } else {
+            return .bothButtons
+        }
+    }
+}
